My First INIT
This is actually my code, Though I script kiddie-ed the Pretty Format Function so that I didn't have to re-invent the wheel.

I like OSDBackground, but it's a dog to configure for each task sequence, and can easily be out dated. 
Also does not function when running from the DP! Wow!
This I my attempt at using the Group Headers to create Dynamic Messages For each task sequence.
Currently relies on tsenv2, as I hate the com load on powershell in WinPEx64, will fix when I finally understand why it sucks so much.